task supports scientific notation

+++ clibs/gameplay/src/core/techtree.cpp
@@ -17,7 +17,7 @@ bool techtree_mgr::is_researched(uint16_t techid) const {
     return researched.contains(techid);
 }
 
-bool techtree_mgr::research_set(uint16_t techid, uint16_t max, uint64_t val) {
+bool techtree_mgr::research_set(uint16_t techid, uint16_t max, uint16_t val) {
     bool finish = false;
     if (val >= max) {
         val = max;
@@ -25,10 +25,10 @@ bool techtree_mgr::research_set(uint16_t techid, uint16_t max, uint64_t val) {
     }
     auto iter = progress.find(techid);
     if (iter == progress.end()) {
-        progress.emplace(techid, (uint16_t)val);
+        progress.emplace(techid, val);
     }
     else {
-        iter->second = (uint16_t)val;
+        iter->second = val;
     }
     if (finish) {
         cache.erase(techid);
